pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-report-to-finmap"
version = "0.1.0"
description = "LangGraph pipeline for invoices → Finmap"
requires-python = ">=3.10"
dependencies = [
  "langgraph",
  "langchain-openai",
  "langchain-community",
  "gspread",
  "google-auth",
  "pydantic>=2",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
